+++ src/discord_client.py
@@ -1,8 +1,9 @@
 from . import curiosa, util
 from .trie import Trie
 
+import asyncio
 import os
-import sys
+import random
 import re
 from asyncio import run as aiorun
 
@@ -15,31 +16,48 @@ load_dotenv()
 DISCORD_BOT_TOKEN = os.getenv('discord-bot-token')
 DISCORD_BOT_MODE = os.getenv('discord-bot-mode')
 
-default_intents = discord.Intents.default()
-default_intents.members = True
-default_intents.message_content = True
+# Status messages chosen by random displayed under the bots name
+STATUS_MESSAGES = [
+    "Brewing magic..",
+    "Casting Blink..",
+    "Sitting by the bonfire..",
+    "Sitting on the Immortal Throne..",
+    "Erupting Vesuvius..",
+    "Pathfinding..",
+    "Navigating the Stormy Seas..",
+    "Attending the Royal Wedding.."
+]
 
-# Maximum length of a message
-MAX_CONTENT_LEN = 1000
+# The time it takes for the status to randomly change in seconds.
+STATUS_UPDATE_TIMER = 300
+
+# Command prefix character, messages starting with this are considered commands.
+COMMAND_PREFIX = '!'
 
 
 class DiscordClient(discord.Client):
     def __init__(self, *, intents=None, **options):
         if intents is None:
-            intents = default_intents
+            intents = discord.Intents.default()
+            intents.members = True
+            intents.message_content = True
 
         super().__init__(intents=intents, **options)
 
         self._browser = None
+        self.current_status = ""
 
-    async def match_command(self, ctx, command):
+    async def handle_command(self, ctx, command):
         """
         Matches commands that start with command prefix to their respective functionality.
         """
+        if command[0] != COMMAND_PREFIX:
+            return
+
         split_command = command.split(" ")
 
-        command = split_command[0][1:len(split_command[0])]
-        parameters = split_command[1:len(split_command)]
+        command = split_command[0][1:]
+        parameters = split_command[1:]
 
         match command:
             case 'deck':
@@ -52,8 +70,32 @@ class DiscordClient(discord.Client):
                 await self.get_card_image(ctx, parameters)
             case c if c in ['faq', 'faqs']:
                 await self.get_faq(ctx, parameters)
+            case 'stop':
+                if DISCORD_BOT_MODE == 'debug':
+                    await self.close_client()
             case _:
                 print(f"Failed to interpret command: {command}")
+
+    async def handle_regex(self, ctx, content):
+        """
+        Handles regex commands
+        """
+        # Regex patterns for commands that are within messages.
+        card_image_pattern = r'\[!(.*?)\]'
+        ci_pattern_match = re.search(card_image_pattern, content)
+
+        card_text_pattern = r'\[\[!(.*?)\]\]'
+        ct_pattern_match = re.search(card_text_pattern, content)
+
+        if ct_pattern_match:
+            card_name = ct_pattern_match.group(1).split(" ")
+            await self.get_card(ctx, card_name)
+            return
+
+        if ci_pattern_match:
+            card_name = ci_pattern_match.group(1).split(" ")
+            await self.get_card_image(ctx, card_name)
+            return
 
     async def get_faq(self, ctx, card_name):
         """
@@ -122,6 +164,7 @@ class DiscordClient(discord.Client):
             self.prefixTree,
             self.cards
         )
+
         await ctx.send(util.code_blockify(received_output))
 
     async def get_card_image(self, ctx, card_name):
@@ -139,49 +182,52 @@ class DiscordClient(discord.Client):
 
         await ctx.send(image_url)
 
+    async def handle_presence(self):
+        """
+        Changes the activity message of bot randomly every 15 minutes
+        """
+        while True:
+            random_status = random.choice(STATUS_MESSAGES)
+            while random_status == self.current_status:
+                random_status = random.choice(STATUS_MESSAGES)
+
+            print(f"Changing presence to: {random_status}")
+            self.current_status = random_status
+
+            await self.change_presence(
+                status=None,
+                activity=CustomActivity(self.current_status)
+            )
+
+            await asyncio.sleep(STATUS_UPDATE_TIMER)
+
     async def on_ready(self):
+        """
+        Handles the event after the client is initialized.
+        """
         print("Archimago now running.")
-        await self.change_presence(status=None, activity=CustomActivity("Brewing magic.."))
+        await self.handle_presence()
 
     async def on_message(self, msg):
+        """
+        Handles the event of message being sent to a server where the client is present.
+        """
         content = msg.content
         channel = msg.channel
 
-        # Don't process messages that are longer than max content length.
-        if len(content) > MAX_CONTENT_LEN or len(content) == 0:
+        # Sanity checks, don't process messages sent by the bot.
+        if len(content) == 0 or msg.author == self.user:
             return
 
-        # Regex patterns for commands that are within messages.
-        card_image_pattern = r'\[!(.*?)\]'
-        ci_pattern_match = re.search(card_image_pattern, content)
-
-        card_text_pattern = r'\[\[!(.*?)\]\]'
-        ct_pattern_match = re.search(card_text_pattern, content)
-
-        # We don't want to receive messages sent by the bot
-        if msg.author == self.user:
-            return
-
-        if "stop" in content and DISCORD_BOT_MODE == 'debug':
-            channel.send("Closing bot client..")
-            curiosa.close_browser()
-            sys.exit()
-
-        if content[0] == "!":
-            await self.match_command(channel, content)
-            return
-
-        if ct_pattern_match:
-            card_name = ct_pattern_match.group(1).split(" ")
-            await self.get_card(channel, card_name)
-            return
-
-        if ci_pattern_match:
-            card_name = ci_pattern_match.group(1).split(" ")
-            await self.get_card_image(channel, card_name)
-            return
+        await self.handle_command(channel, content)
+        await self.handle_regex(channel, content)
 
     def start_client(self, browser):
+        """
+        Starts the discord client and initializes card data.
+
+        If card data fails to load, throws an exception.
+        """
         if browser is None:
             print("Selenium webdriver required for starting discord client.")
             return
@@ -197,8 +243,13 @@ class DiscordClient(discord.Client):
         try:
             aiorun(runner())
         except KeyboardInterrupt:
+            self._browser.quit()
             return
 
     async def close_client(self):
+        """
+        Closes the discord client instance.
+        """
+        print("Closing Archimago..")
         self._browser.close()
         await self.close()